Moves BannerText to sections folder
