+++ src/components/sections/BannerText/BannerText.tsx
