+++ src/components/sections/BannerText/index.tsx
